Rename output writers to file handlers
